% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{as.df}
\alias{as.df}
\title{as.df shortcut function}
\usage{
as.df(x)
}
\arguments{
\item{x}{object to convert to data.frame without factors}
}
\value{
df
}
\description{
as.df shortcut function
}
\examples{
x = as.df(c(1,2,3))
}
