% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_ct_data.R
\name{boost_ct_data}
\alias{boost_ct_data}
\title{Enhance a dataframe by searching the clinicaltrials API for values in a specified column and requesting additional fields of results from the API to add as new columns to your data.frame}
\usage{
boost_ct_data(df, search_col_name_str, request_fields, api_terms_max = 150)
}
\arguments{
\item{df}{starting dataframe}

\item{search_col_name_str}{character string naming the column with values to search for}

\item{request_fields}{vector of field names to request from the API}

\item{api_terms_max}{default 150, but reduce this is the fields are data-heavy and the API throws an error}
}
\value{
dataframe plus result fields from the API
}
\description{
Enhance a dataframe by searching the clinicaltrials API for values in a specified column and requesting additional fields of results from the API to add as new columns to your data.frame
}
