% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{eval_as_text}
\alias{eval_as_text}
\title{take a variable input as text and parse/eval it}
\usage{
eval_as_text(x)
}
\arguments{
\item{x}{input value}
}
\value{

}
\description{
take a variable input as text and parse/eval it
}
