% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ct_feature_cols.R
\name{expand_ct_feature_cols}
\alias{expand_ct_feature_cols}
\title{Unnest_longer the vector values for sets of columns related to condition and intervention that should remain correlated}
\usage{
expand_ct_feature_cols(df, features = c("Intervention", "Condition"))
}
\arguments{
\item{df}{the input data.frame}

\item{features}{either "Intervention" or "Condition". It may be necessary to cut long dfs to fewer rows if the amount of data would exceed ram limits}
}
\value{
a longer dataframe
}
\description{
Unnest_longer the vector values for sets of columns related to condition and intervention that should remain correlated
}
\examples{
df = df \%>\%

}
