% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_data.R
\name{get_ct_data}
\alias{get_ct_data}
\title{request result fields from clinicaltrials API}
\usage{
get_ct_data(
  search_expr,
  request_fields = get_ct_fields(),
  expr_in_field = NULL,
  count = NULL,
  remove_null_results = F
)
}
\arguments{
\item{search_expr}{term to search for}

\item{request_fields}{names of the fields to request from the API}

\item{expr_in_field}{field where the search expression must be found}

\item{count}{developer note: can't remember why this param is here}

\item{remove_null_results}{should fields with nothing but null values be removed automatically from the result df?}
}
\value{
results as a data.frame
}
\description{
request result fields from clinicaltrials API
}
\examples{
df = get_ct_data("myalgic encephalomyelitis", request_fields = c("NCTId", get_ct_fields(starts_with_str = "Condition")))

}
