% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_fields.R
\name{get_ct_fields}
\alias{get_ct_fields}
\title{request a vector of the names of fields available from the clinicaltrials API}
\usage{
get_ct_fields(
  all = F,
  starts_with_str = NULL,
  contains_str = NULL,
  case_sensitive = F
)
}
\arguments{
\item{all}{developer's note: this may be a leftover from previous code - ignore it}

\item{starts_with_str}{get fields that start with this character substring}

\item{contains_str}{get fields that contain this character substring}

\item{case_sensitive}{is it case sensitive?}
}
\value{
a vector naming the fields available via the ClinicalTrials API, optionally filtered
}
\description{
request a vector of the names of fields available from the clinicaltrials API
}
