% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{no_na_cols}
\alias{no_na_cols}
\title{Remove columns from a data.frame that have only NA values}
\usage{
no_na_cols(df)
}
\arguments{
\item{df}{input data.frame}
}
\value{

}
\description{
Remove columns from a data.frame that have only NA values
}
\examples{
df = data.frame(first = c(1,2,3,NA), second = c("a","b","c",NA), third = c(NA,NA,NA,NA)) \%>\% no_na_cols()
}
