% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{prefix}
\alias{prefix}
\title{paste a prefix before value or values}
\usage{
prefix(x, prefix)
}
\arguments{
\item{x}{input value}

\item{prefix}{character string to add as prefix}
}
\value{

}
\description{
paste a prefix before value or values
}
