% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{round_up}
\alias{round_up}
\title{round up to a number of digits}
\usage{
round_up(x, digits = 0)
}
\arguments{
\item{x}{input value}

\item{digits}{number of digits for rounding precision}
}
\value{

}
\description{
round up to a number of digits
}
