% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{split_longer_any_vector_cols}
\alias{split_longer_any_vector_cols}
\title{split and unnest_longer all the columns of a df that contain vectors posing as text values}
\usage{
split_longer_any_vector_cols(df)
}
\arguments{
\item{df}{input data.frame}
}
\value{

}
\description{
split and unnest_longer all the columns of a df that contain vectors posing as text values
}
\examples{
df = get_ct_data("myalgic encephalomyelitis", request_fields = c("NCTId", "Condition")) \%>\% split_longer_any_vector_cols()

}
