% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{split_longer_vector}
\alias{split_longer_vector}
\title{split and unnest_longer the values in a column that are vectors posing as text}
\usage{
split_longer_vector(df, col_name_chr)
}
\arguments{
\item{df}{input dataframe}

\item{col_name_chr}{name of the column to split}
}
\value{

}
\description{
split and unnest_longer the values in a column that are vectors posing as text
}
