% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{substrRight}
\alias{substrRight}
\title{get the last n characters from the end of a string}
\usage{
substrRight(x, n)
}
\arguments{
\item{n}{}
}
\value{
a character substring
}
\description{
get the last n characters from the end of a string
}
\examples{
substrRight("malfunction", 8)

}
