% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{unvector}
\alias{unvector}
\title{convert vector-like text into a real vector}
\usage{
unvector(x)
}
\arguments{
\item{x}{input value}
}
\value{

}
\description{
convert vector-like text into a real vector
}
\examples{
unvector("c(\"x\", \"y\")")

}
