% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_data.R
\name{ct_search}
\alias{ct_search}
\title{search for a term in clinicaltrials data}
\usage{
ct_search(
  search_expr = "fatigue",
  request_fields = "NCTId,Condition,BriefTitle,OrgStudyId,EligibilityCriteria",
  count = NULL
)
}
\arguments{
\item{search_expr}{the expression to search for}

\item{request_fields}{which fields to request from the API}

\item{count}{developer's note: can't remember why this is here... ignore it}
}
\value{
results as a data.frame
}
\description{
search for a term in clinicaltrials data
}
