% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_convenience_functions.R
\name{split_longer}
\alias{split_longer}
\title{split and unnest_longer a column in a df by some regex value}
\usage{
split_longer(df, col_to_split, by_regex, encoding = "UTF-8")
}
\arguments{
\item{df}{input data.frame}

\item{col_to_split}{the column to split}

\item{by_regex}{regex used as the splitter}

\item{encoding}{character encoding}
}
\value{

}
\description{
split and unnest_longer a column in a df by some regex value
}
